refactor(test, query): use integration tests
makes the code easier to edit + simplifies test creation :)
